Defer PVE & PVT loader code to post-load time
