+++ src/features/page_view_event/instrument/index.js
@@ -1,7 +1,6 @@
 
 import { handle } from '../../../common/event-emitter/handle'
 import { isiOS } from '../../../common/util/user-agent'
-import { onTTFB } from 'web-vitals'
 import { InstrumentBase } from '../../utils/instrument-base'
 import * as CONSTANTS from '../constants'
 import { FEATURE_NAMES } from '../../../loaders/features/features'
@@ -13,30 +12,17 @@ export class Instrument extends InstrumentBase {
   static featureName = CONSTANTS.FEATURE_NAME
   constructor (agentIdentifier, aggregator, auto = true) {
     super(agentIdentifier, aggregator, CONSTANTS.FEATURE_NAME, auto)
-    const agentRuntime = getRuntime(agentIdentifier) // we'll store timing values on the runtime obj to be read by the aggregate module
 
     if (typeof PerformanceNavigationTiming !== 'undefined' && !isiOS) {
-      this.alreadySent = false // we don't support timings on BFCache restores
-
-      /* Time To First Byte
-        This listener subscribes to the window load event and must record these values *before* PVE's aggregate sends RUM. */
-      onTTFB(({ value, entries }) => {
-        if (this.alreadySent) return
-        this.alreadySent = true
-
-        agentRuntime[CONSTANTS.TTFB] = Math.round(value) // this is our "backend" duration; web-vitals will ensure it's lower bounded at 0
-
-        // Similar to what vitals does for ttfb, we have to factor in activation-start when calculating relative timings:
-        const navEntry = entries[0]
-        const respOrActivStart = Math.max(navEntry.responseStart, navEntry.activationStart || 0)
-        agentRuntime[CONSTANTS.FBTWL] = Math.max(Math.round(navEntry.loadEventEnd - respOrActivStart), 0) // our "frontend" duration
-        handle('timing', ['load', Math.round(navEntry.loadEventEnd)], undefined, FEATURE_NAMES.pageViewTiming, this.ee)
-        agentRuntime[CONSTANTS.FBTDC] = Math.max(Math.round(navEntry.domContentLoadedEventEnd - respOrActivStart), 0) // our "dom processing" duration
-      })
+      /* This block has been moved to the aggregate portion which is fetched and executed after page
+          load in an effort to reduce loader bundle size.
+      */
     } else if (typeof PerformanceTiming !== 'undefined') {
       // This is the predecessor to PerfNavTiming, needed because that isn't supported on pre v15 safari and v15.2 ios, which are problematic with onTTFB using that API.
       // Timings here are calculated on a best-effort basis.
       // *cli Mar'23 - iOS 15.2 & 15.4 testing in Sauce still fails with onTTFB. Hence, all iOS will fallback to this for now.
+      const agentRuntime = getRuntime(agentIdentifier)
+
       agentRuntime[CONSTANTS.TTFB] = Math.max(Date.now() - agentRuntime.offset, 0)
       onDOMContentLoaded(() => agentRuntime[CONSTANTS.FBTDC] = Math.max(now() - agentRuntime[CONSTANTS.TTFB], 0))
       onWindowLoad(() => {
