+++ src/features/page_view_timing/first-paint.js
