+++ src/features/page_view_timing/instrument/index.js
@@ -3,16 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 import { handle } from '../../../common/event-emitter/handle'
-import { onFCP, onFID, onLCP, onCLS, onINP } from 'web-vitals'
-import { onFirstPaint } from './first-paint'
-import { onLongTask } from './long-tasks'
+import { getRuntime } from '../../../common/config/config'
 import { subscribeToVisibilityChange } from '../../../common/window/page-visibility'
 import { windowAddEventListener } from '../../../common/event-listener/event-listener-opts'
 import { now } from '../../../common/timing/now'
-import { getConfigurationValue } from '../../../common/config/config'
 import { InstrumentBase } from '../../utils/instrument-base'
 import { FEATURE_NAME } from '../constants'
-import { FEATURE_NAMES } from '../../../loaders/features/features'
 import { isBrowserScope } from '../../../common/util/global-scope'
 
 export class Instrument extends InstrumentBase {
@@ -21,92 +17,14 @@ export class Instrument extends InstrumentBase {
     super(agentIdentifier, aggregator, FEATURE_NAME, auto)
     if (!isBrowserScope) return // CWV is irrelevant outside web context
 
-    // Since we don't support timings on BFCache restores, this tracks and helps cap metrics that web-vitals report more than once.
-    this.alreadySent = new Set()
+    // Document visibility state becomes hidden; this should run as soon as possible in page life.
+    // While we try to replicate web-vital's visibilitywatcher logic in an effort to defer that library to post-pageload, this isn't perfect and doesn't consider prerendering.
+    getRuntime(agentIdentifier).initHidden = Boolean(document.visibilityState === 'hidden')
+    subscribeToVisibilityChange(() => handle('docHidden', [now()], undefined, FEATURE_NAME, this.ee), true)
 
-    /* PerformancePaintTiming API
-        This listener is deferrable BUT because it does not consider document vis state yet, unlike FID and LCP below. BFC is also not supported. */
-    onFirstPaint(({ name, value }) => {
-      handle('timing', [name.toLowerCase(), Math.floor(value)], undefined, FEATURE_NAMES.pageViewTiming, this.ee)
-    })
-
-    /* First Contentful Paint
-        This listener cannot be deferred yet maintain full functionality under v3. Reason: it relies on detecting document vis state asap from time origin. */
-    onFCP(({ name, value }) => {
-      if (this.alreadySent.has(name)) return
-      this.alreadySent.add(name)
-
-      handle('timing', [name.toLowerCase(), value], undefined, FEATURE_NAMES.pageViewTiming, this.ee)
-    })
-
-    /* First Input Delay (under "First Interaction")
-        This listener cannot be deferred yet maintain full functionality under v3. Reason: it relies on detecting document vis state asap from time origin. */
-    onFID(({ name, value, entries }) => {
-      if (this.alreadySent.has(name)) return
-      this.alreadySent.add(name)
-
-      // CWV will only report one (THE) first-input entry to us; fid isn't reported if there are no user interactions occurs before the *first* page hiding.
-      const fiEntry = entries[0]
-      const attributes = {
-        type: fiEntry.name,
-        fid: Math.round(value)
-      }
-      this.addConnectionAttributes(attributes)
-      handle('timing', ['fi', Math.round(fiEntry.startTime), attributes], undefined, FEATURE_NAMES.pageViewTiming, this.ee)
-    })
-
-    /* Largest Contentful Paint
-        This listener cannot be deferred yet maintain full functionality under v3. Reason: it relies on detecting document vis state asap from time origin. */
-    onLCP(({ name, value, entries }) => {
-      if (this.alreadySent.has(name)) return
-      this.alreadySent.add(name)
-
-      // CWV will only ever report one (THE) lcp entry to us; lcp is also only reported *once* on earlier(user interaction, page hidden).
-      const lcpEntry = entries[entries.length - 1] // this looks weird if we only expect one, but this is how cwv-attribution gets it so to be sure...
-      const attributes = this.addConnectionAttributes({})
-      handle('lcp', [value, lcpEntry, attributes], undefined, FEATURE_NAMES.pageViewTiming, this.ee)
-    })
-
-    /* Cumulative Layout Shift
-        This listener IS deferrable, though further validation required. */
-    onCLS(({ value }) => {
-      // We don't have to limit this callback since cls is stored as state var and only sent as attribute on other timings.
-      handle('cls', [value], undefined, FEATURE_NAMES.pageViewTiming, this.ee)
-    })
-
-    /* Interaction-to-Next-Paint
-        This listener IS deferrable, though further validation required. */
-    onINP(({ name, value, id }) => {
-      handle('timing', [name.toLowerCase(), value, { metricId: id }], undefined, FEATURE_NAMES.pageViewTiming, this.ee)
-    })
-
-    /* PerformanceLongTaskTiming API
-        This listener IS deferrable. */
-    if (getConfigurationValue(this.agentIdentifier, 'page_view_timing.long_task') === true) {
-      onLongTask(({ name, value, info }) => {
-        handle('timing', [name.toLowerCase(), value, info], undefined, FEATURE_NAMES.pageViewTiming, this.ee) // lt context is passed as attrs in the timing node
-      })
-    }
-
-    // Document visibility state becomes hidden; cannot assume safely deferrable.
-    subscribeToVisibilityChange(() => handle('docHidden', [now()], undefined, FEATURE_NAMES.pageViewTiming, this.ee), true)
-
-    // Window fires its pagehide event (typically on navigation--this occurrence is a *subset* of vis change); cannot assume safely deferrable.
-    windowAddEventListener('pagehide', () => handle('winPagehide', [now()], undefined, FEATURE_NAMES.pageViewTiming, this.ee))
+    // Window fires its pagehide event (typically on navigation--this occurrence is a *subset* of vis change); don't defer this unless it's guarantee it cannot happen before load(?)
+    windowAddEventListener('pagehide', () => handle('winPagehide', [now()], undefined, FEATURE_NAME, this.ee))
 
     this.importAggregator()
-  }
-
-  // takes an attributes object and appends connection attributes if available
-  addConnectionAttributes (attributes) {
-    var connection = navigator.connection || navigator.mozConnection || navigator.webkitConnection // to date, both window & worker shares the same support for connection
-    if (!connection) return
-
-    if (connection.type) attributes['net-type'] = connection.type
-    if (connection.effectiveType) attributes['net-etype'] = connection.effectiveType
-    if (connection.rtt) attributes['net-rtt'] = connection.rtt
-    if (connection.downlink) attributes['net-dlink'] = connection.downlink
-
-    return attributes
   }
 }
